updated 04_more_exercises

+++ 04_more_exercises/data_types.py
@@ -3,7 +3,7 @@ def find_the_data_type(type, operation):
         result = int(operation) * 2
         print(result)
     elif type == 'real':
-        result = int(operation) * 1.5
+        result = float(operation) * 1.5
         print(f'{result:.2f}')
     elif type == 'string':
         print(f'${operation}$')
@@ -13,3 +13,4 @@ data_type = input()
 activity = input()
 
 find_the_data_type(data_type, activity)
+
